requirement.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
deprecation==2.1.0
exceptiongroup==1.2.0
Faker==24.9.0
fastapi==0.110.1
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
google-pasta==0.2.0
gotrue==2.4.2
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
joblib==1.3.2
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
pillow==10.3.0
postgrest==0.16.2
protobuf==4.25.3
psycopg-binary==3.1.18
PuLP==2.8.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
realtime==1.0.3
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sklearn==0.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.1
supafunc==0.4.5
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.4.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
